requirements_windows.txt
Kivy
kivy-deps.glew
kivy-deps.gstreamer
kivy-deps.sdl2
Kivy-Garden
japanize-kivy
numpy
pyserial
pillow
scipy
pywin32
